REFACTOR: User <-> Interest 리팩토링 (#24)
