--- src/main/java/com/deadlock/hellocs/domain/interest/entity/Interest.java
@@ -1,9 +1,0 @@
-package com.deadlock.hellocs.domain.interest.entity;
-
-public enum Interest {
-    FRONTEND,
-    BACKEND,
-    INFRA,
-    ALGORITHM,
-    AI
-}
