--- src/main/java/com/deadlock/hellocs/domain/interest/entity/UserInterestId.java
@@ -1,24 +1,0 @@
-package com.deadlock.hellocs.domain.interest.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@Embeddable
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode
-public class UserInterestId implements Serializable {
-
-    @Column(name = "user_id")
-    private Long userId;
-
-    @Column(name = "interest_id")
-    private Long interestId;
-}
