+++ src/main/java/com/deadlock/hellocs/domain/user/entity/UserInterest.java
@@ -1,0 +1,20 @@
+package com.deadlock.hellocs.domain.user.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+
+import java.util.List;
+
+@Entity
+@Getter
+public class UserInterest {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "user_id", nullable = false)
+    private Long userId;
+
+    @Column(name = "topic_id", nullable = false)
+    private Long topicId;
+}
